+++ src/main/java/me/partlysunny/regionous/statics/StaticSphereRegion.java
@@ -7,7 +7,7 @@ public class StaticSphereRegion extends SphereRegion {
 
     private final Location location;
 
-    public StaticSphereRegion(String identifier, float radius, Location location) {
+    public StaticSphereRegion(String identifier, double radius, Location location) {
         super(identifier, radius);
         this.location = location;
     }
